✨Feat: Implement business logic, use cases and unit tests for order domain
- OrderController 의 주문 생성 API (`createOrder`), 특정 주문 조회 API (`getOrder`), 주문 상태 API (`updateOrderStatus`) 구현
- application 레이어에 CreateOrderUseCase, GetOrderUseCase, UpdateOrderStatusUseCase 구현
- domain 레이어에 OrderService 필요 기능 구현
- OrderService 및 연관 도메인에 대하여 단위 테스트 추가

Resolved #7, #9, #28

+++ src/main/java/com/jane/ecommerce/interfaces/dto/order/OrderResponse.java
@@ -17,6 +17,6 @@ public class OrderResponse {
     String id;
     String status;
     List<OrderItemDTO> orderItems;
-    int totalAmount;
+    Long totalAmount;
     LocalDateTime createAt;
 }
